pyproject.toml
[project]
name = "mcp-doc-indexer"
version = "0.1.0"
description = "MCP server for local document indexing and search using LanceDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lancedb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "watchdog>=3.0.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "fastmcp>=0.1.0",
    "ollama>=0.1.7",
    "pydantic>=2.0.0",
    "asyncio",
    "numpy",
    "pandas",
    "tiktoken",
    "python-dotenv"
]

[project.scripts]
mcp-doc-indexer = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]